Add Photo tests and Theme's route

+++ server/src/controllers/MemberController.js
@@ -38,6 +38,11 @@ module.exports = {
   async remove(req, res) {
     const { memberId } = req.params;
 
+    const member = await Team.findByPk(memberId);
+    if (!member) {
+      return res.status(400).json({ error: `O Tema ${memberId} não existe` })
+    }
+
     await Member.destroy({
       where: {
         id: memberId
